+++ drivers/adc/CMakeLists.txt
@@ -1,8 +1,8 @@
 zephyr_library()
 
-zephyr_library_sources_ifdef(CONFIG_ADC_DW		adc_dw.c)
 zephyr_library_sources_ifdef(CONFIG_ADC_MCUX_ADC16	adc_mcux_adc16.c)
 zephyr_library_sources_ifdef(CONFIG_ADC_SAM_AFEC	adc_sam_afec.c)
 zephyr_library_sources_ifdef(CONFIG_ADC_NRFX_ADC	adc_nrfx_adc.c)
 zephyr_library_sources_ifdef(CONFIG_ADC_NRFX_SAADC	adc_nrfx_saadc.c)
+zephyr_library_sources_ifdef(CONFIG_ADC_INTEL_QUARK_SE_C1000_SS	adc_intel_quark_se_c1000_ss.c)
 zephyr_library_sources_ifdef(CONFIG_ADC_INTEL_QUARK_D2000	adc_intel_quark_d2000.c)
