+++ drivers/adc/Kconfig
@@ -47,8 +47,6 @@ config ADC_0
 config ADC_1
 	bool "Enable ADC 1"
 
-source "drivers/adc/Kconfig.dw"
-
 source "drivers/adc/Kconfig.mcux"
 
 source "drivers/adc/Kconfig.nrfx"
