--- drivers/adc/Kconfig.dw
@@ -1,71 +1,0 @@
-# Kconfig - ADC configuration options
-
-#
-# Copyright (c) 2015 Intel Corporation
-#
-# SPDX-License-Identifier: Apache-2.0
-#
-
-menuconfig ADC_DW
-	bool "ARC Designware Driver"
-	depends on ARC
-	select ADC_0
-	help
-	  Enable the driver implementation of the Designware ADC IP.
-
-if ADC_DW
-
-
-config ADC_DW_CALIBRATION
-	bool "Enable Calibration"
-	default y
-	help
-	  Enables ADC to run with a calibrated output at the
-	  expense of execution speed when exiting low power states.
-	  If disabled, the ADC will require the application/system-integrator
-	  to provide a calibration method.
-
-choice
-	prompt "Output Mode"
-	default ADC_DW_SERIAL
-	help
-	  ADC output mode: parallel or serial.
-
-config ADC_DW_SERIAL
-	bool "Serial"
-
-config ADC_DW_PARALLEL
-	bool "Parallel"
-
-endchoice
-
-choice
-	prompt "Capture Mode"
-	default ADC_DW_RISING_EDGE
-	help
-	  ADC controller capture mode:
-	  by rising or falling edge of adc_clk
-
-config ADC_DW_RISING_EDGE
-	bool "Rising Edge"
-
-config ADC_DW_FALLING_EDGE
-	bool "Falling Edge"
-
-endchoice
-
-config ADC_DW_SERIAL_DELAY
-	int "Serial Delay"
-	default 1
-	help
-	  Number of ADC clock ticks that the first bit of
-	  the serial output is delayed after the conversion
-	  has started.
-
-config ADC_DW_CLOCK_RATIO
-	int "Clock Ratio"
-	default 1024
-	help
-	  ADC Clock Ratio
-
-endif # ADC_DW
